removal of special machine edge types that are not needed these days
