+++ spynnaker/pyNN/extra_algorithms/splitter_components/splitter_abstract_pop_vertex_slice.py
@@ -69,16 +69,12 @@ class SplitterAbstractPopulationVertexSlice(
 
     @overrides(AbstractSplitterSlice.get_pre_vertices)
     def get_pre_vertices(self, edge, outgoing_edge_partition):
-        extra_pre_edge_types = [MachineEdge]
-        extra_pre_edge_types.extend(self.extra_pre_edge_type())
-        return self._get_map(extra_pre_edge_types)
+        return self._get_map([MachineEdge])
 
     @overrides(AbstractSplitterSlice.get_post_vertices)
     def get_post_vertices(
             self, edge, outgoing_edge_partition, src_machine_vertex):
-        extra_pre_edge_types = [MachineEdge]
-        extra_pre_edge_types.extend(self.extra_post_edge_type())
-        return self._get_map(extra_pre_edge_types)
+        return self._get_map([MachineEdge])
 
     @overrides(AbstractSplitterSlice.create_machine_vertex)
     def create_machine_vertex(
