+++ spynnaker/pyNN/extra_algorithms/splitter_components/spynnaker_splitter_slice_legacy.py
@@ -28,13 +28,9 @@ class SpynnakerSplitterSliceLegacy(
 
     @overrides(SplitterSliceLegacy.get_pre_vertices)
     def get_pre_vertices(self, edge, outgoing_edge_partition):
-        extra_pre_edge_types = [MachineEdge]
-        extra_pre_edge_types.extend(self.extra_pre_edge_type())
-        return self._get_map(extra_pre_edge_types)
+        return self._get_map([MachineEdge])
 
     @overrides(SplitterSliceLegacy.get_post_vertices)
     def get_post_vertices(
             self, edge, outgoing_edge_partition, src_machine_vertex):
-        extra_pre_edge_types = [MachineEdge]
-        extra_pre_edge_types.extend(self.extra_post_edge_type())
-        return self._get_map(extra_pre_edge_types)
+        return self._get_map([MachineEdge])
